Bug #34911: fixed forcing of new access settings for shares
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-3.2/ucs-school-3.2r2@50532 1b283449-9f2f-0410-b571-8eb228e1a901

+++ ucs-school-umc-computerroom/umc/python/computerroom/__init__.py
@@ -809,11 +809,11 @@ class Instance(SchoolBaseModule):
 
 		# reset SMB connections
 		smbstatus = SMB_Status()
-		italc_users = self._italc.users
+		italc_users = [x.lower() for x in self._italc.users if x]
 		MODULE.info('iTALC users: %s' % ', '.join(italc_users))
 		for process in smbstatus:
 			MODULE.info('SMB process: %s' % str(process))
-			if process.username in italc_users:
+			if process.username and process.username.lower() in italc_users:
 				MODULE.info('Kill SMB process %s' % process.pid)
 				os.kill(int(process.pid), signal.SIGTERM)
 		_finished()
